requirements.txt
# The Hive - AI Agent Ecosystem
# Python Dependencies
# Generated: 2025-11-17

# Core Framework
flask==3.0.0
python-dotenv==1.0.0

# AI APIs
groq==0.4.1                    # Fast inference (Groq Cloud)
openai==1.6.1                  # Compatibility with local models
google-generativeai==0.3.2     # Gemini API (FREE tier)
langchain==0.1.0               # LLM orchestration (optional)
langchain-community==0.0.13    # Community integrations

# Local LLM (DeepSeek via Ollama)
ollama==0.1.6                  # Ollama Python client
llama-cpp-python==0.2.27       # Alternative: LlamaCpp direct

# Vector Database & Memory
chromadb==0.4.22               # Local vector DB
sentence-transformers==2.2.2   # Embeddings

# Web Scraping & Research (para Niche Discovery)
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0               # For JS-heavy sites
playwright==1.40.0             # Alternative to Selenium

# WordPress Integration (si Congress lo descubre útil)
python-wordpress-xmlrpc==2.3   # WordPress API client

# Automation (opcional - según descubrimientos autónomos)
Appium-Python-Client==3.1.0    # Mobile automation
uiautomator2==3.0.0            # Android automation

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
pydantic==2.5.3                # Data validation
PyYAML==6.0.1                  # Config files
click==8.1.7                   # CLI interface
rich==13.7.0                   # Beautiful terminal output
tqdm==4.66.1                   # Progress bars
psutil==5.9.6                  # System monitoring (for workers)

# Blockchain & Economy
web3==6.11.3                   # Web3 integration (BSC)

# Telegram Bot (Leo's oversight interface)
python-telegram-bot==20.7      # Telegram Bot API with async support
eth-account==0.9.0             # Ethereum account management
py-solc-x==2.0.2               # Solidity compiler
cryptography==41.0.7           # Encryption for laws

# Testing
pytest==7.4.3
pytest-cov==4.1.0              # Coverage reports
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1                 # Code formatter
flake8==7.0.0                  # Linter
mypy==1.7.1                    # Type checker
isort==5.13.2                  # Import sorter

# Monitoring & Metrics
prometheus-client==0.19.0      # Metrics collection
psutil==5.9.6                  # System monitoring

# Async Support
aiohttp==3.9.1                 # Async HTTP
asyncio==3.4.3

# Security
cryptography==41.0.7
